A fix to more problems than I can count
Specifically Google's wierd constants
Also moved sources out of main and increased query rate
