+++ main.py
@@ -10,8 +10,11 @@ except ImportError:
 	import urllib.request as urllib2
 	from urllib.parse import urljoin
 
+from google.cloud.language import enums
+
 from news import *
 from annotator import Annotator
+from sources import AlJazeera, Bbc, Guardian, Hill, Wapo
 
 reddit = praw.Reddit('sourcesbot', user_agent='web:sourcesbot:v0.0.2 by /u/michaelh115')
 
@@ -73,261 +76,28 @@ def get_characteristics_from_list(trees, r):
 def get_characteristics(t, r):
 	# Get people, places, organizations, actions and other things of signifigance
 	if t.entity is not None:
-		if t.entity.type == 'LOCATION':
+		if t.entity.type == 2: # A location
 			if t.syntax.simplified not in r['talley']['places']:
 				r['talley']['places'].append(t.syntax.simplified)
-		elif t.entity.type == 'ORGANIZATION':
+		elif t.entity.type == 3: # An organization
 			if t.syntax.simplified not in r['talley']['organizations']:
 				r['talley']['organizations'].append(t.syntax.simplified)
-		elif t.entity.type == 'PERSON':
+		elif t.entity.type == 1: # A person
 			if t.syntax.simplified not in r['talley']['people']:
 				r['talley']['people'].append(t.syntax.simplified)
 	else:
-		if t.syntax.tag == 'NOUN':
+		if t.syntax.tag == 6: # A noun
 			if t.syntax.simplified not in r['talley']['things']:
-						r['talley']['things'].append(t.syntax.simplified)
-		elif t.syntax.tag == 'VERB':
+				r['talley']['things'].append(t.syntax.simplified)
+		elif t.syntax.tag == 11: # A verb (need to figure out how to import constants)
 			if t.syntax.simplified not in r['talley']['actions']:
-						r['talley']['actions'].append(t.syntax.simplified)
+				r['talley']['actions'].append(t.syntax.simplified)
 		# needs to be rewritten
 		#elif t.syntax.tag == 'PUNCT':
 		#	r = find_mandatory_words_end(t, n, r)
 	for b in t.branches:
 		get_characteristics(b, r)
 	return(r)
-
-class Source:
-	def __init__(self):
-		self._time = None
-		self._content = None
-
-	def get(self):
-		refresh = datetime.timedelta(minutes=30)
-		if datetime.datetime.utcnow() - self._time > refresh:
-			self.update()
-		return(self._content)
-
-	def update(self):
-		raise NotImplementedError("Must override update")
-
-	def _characterize(self):
-		for s in range(len(self._content)):
-			# The default r NEEDS to be passed or the function below becomes possessed
-			c = get_characteristics_from_list(
-				self._content[s]['machine_title'],
-				{
-					'talley':{
-						'places':[],
-						'people':[],
-						'organizations':[],
-						'things':[],
-						'actions':[]
-					},
-					'mandate':{
-						'speakers':[]
-					}
-				}
-			)
-			c = get_characteristics_from_list(self._content[s]['description'], c)
-			c = dedup_entities(c)
-			self._content[s]['character'] = c
-
-class AlJazeera(Source):
-	def __isolate_content(self, links):
-		self._content = []
-		for h in links:
-			desc, title = '', ''
-			try:
-				if str(h.contents[0]).startswith('<'):
-					if not str(h.contents[0].contents[0]).startswith('<'):
-						title = h.contents[0].contents[0]
-						try:
-							if 'top-sec-desc' in h.parent.contents[3]['class']:
-								desc = h.parent.contents[3].contents[0]
-						except KeyError:
-							pass
-			except IndexError: # malformed HTML
-				pass
-			try:
-				url = h['href']
-				if url.startswith('/'):
-					url = urljoin('http://www.aljazeera.com/', url)
-			except KeyError: # Yes, here at Al Jazeera we use empty <a> tags!
-				pass
-			machine_title = annotator.annotate(title)
-			thresh = 2
-			if desc:
-				thresh += 1
-			desc = annotator.annotate(desc)
-			self._content.append({
-				'title':title,
-				'machine_title':machine_title,
-				'description':desc,
-				'threshold':thresh,
-				'url':url
-			})
-
-	def update(self):
-		self._time = datetime.datetime.utcnow()
-		r = urllib2.urlopen("http://www.aljazeera.com/")
-		html = r.read()
-		soup = BeautifulSoup(html, "html.parser")
-		links = soup.find_all('a')
-		self.__isolate_content(links)
-		self._characterize()
-
-class Bbc(Source):
-	def __isolate_content(self, stories):
-		self._content = []
-		for s in stories:
-			desc = ''
-			h = s.find_all('a', {'class':'gs-c-promo-heading'})
-			if len(h) > 0:
-				if not str(h[0].contents[0].contents[0]).startswith('<'):
-					title = h[0].contents[0].contents[0]
-				else:
-					title = h[0].contents[1].contents[0]
-				d = s.find_all({'class':'gs-c-promo-summary'})
-				if len(d) > 0:
-						desc = d[0].contents[0]
-				url = h[0]['href']
-				if url.startswith('/'):
-					url = urljoin('http://www.bbc.com/news', url)
-				machine_title = annotator.annotate(title)
-				thresh = 2
-				if desc:
-					thresh += 1
-				desc = annotator.annotate(desc)
-				self._content.append({
-					'title':title,
-					'machine_title':machine_title,
-					'description':desc,
-					'threshold':thresh,
-					'url':url
-				})
-
-	def update(self):
-		self._time = datetime.datetime.utcnow()
-		r = urllib2.urlopen("http://www.bbc.com/news")
-		html = r.read()
-		soup = BeautifulSoup(html, "html.parser")
-		links = soup.find_all('div', {'class':'gs-c-promo'})
-		self.__isolate_content(links)
-		self._characterize()
-
-class Guardian(Source):
-	def __isolate_content(self, links):
-		self._content = []
-		for h in links:
-			desc = ''
-			try:
-				title = h.contents[0]
-			except TypeError:
-				if str(h.contents[0].contents[0]).startswith('<'):
-					title = h.contents[1].contents[0]
-				else:
-					title = h.contents[0].contents[0]
-			url = h['href']
-			if url.startswith('/'):
-				url = urljoin('https://www.theguardian.com', url)
-			machine_title = annotator.annotate(title)
-			thresh = 2
-			if desc:
-				thresh += 1
-			desc = annotator.annotate(desc)
-			self._content.append({
-				'title':title,
-				'machine_title':machine_title,
-				'description':desc,
-				'threshold':thresh,
-				'url':url
-			})
-
-	def update(self):
-		self._time = datetime.datetime.utcnow()
-		r = urllib2.urlopen("https://www.theguardian.com")
-		html = r.read()
-		soup = BeautifulSoup(html, "html.parser")
-		links = soup.find_all('a', {'class':'js-headline-text'})
-		self.__isolate_content(links)
-		self._characterize()
-
-class Hill(Source):
-	def __isolate_content(self, links):
-		self._content = []
-		for h in links:
-			desc = ''
-			title = ''
-			try:
-				if ('menu__link' not in h['class'] and 
-					'hide_overlay' not in h['class']):
-					if not h['class'][0].startswith('social-share-'):
-						try:
-							title = h.contents[0]
-						except IndexError:
-							pass
-			except KeyError:
-				try:
-					title = h.contents[0]
-				except IndexError:
-					pass
-			url = h['href']
-			if url.startswith('/'):
-				url = urljoin('http://thehill.com/', url)
-			machine_title = annotator.annotate(title)
-			thresh = 2
-			if desc:
-				thresh += 1
-			desc = annotator.annotate(desc)
-			self._content.append({
-				'title':title,
-				'machine_title':machine_title,
-				'description':desc,
-				'threshold':thresh,
-				'url':url
-			})
-
-	def update(self):
-		self._time = datetime.datetime.utcnow()
-		r = urllib2.urlopen("http://thehill.com/")
-		html = r.read()
-		soup = BeautifulSoup(html, "html.parser")
-		links = soup.find_all('a')
-		self.__isolate_content(links)
-		self._characterize()
-
-class Wapo(Source):
-	def __isolate_content(self, links):
-		self._content = []
-		for h in links:
-			desc = ''
-			title = h.contents[0]
-			if 'blurb' in h.parent.parent.contents[1]['class']:
-				desc = h.parent.parent.contents[1].contents[0]
-			url = h['href']
-			if url.startswith('/'):
-				url = urljoin('http://www.washingtonpost.com', url)
-			machine_title = annotator.annotate(title)
-			thresh = 2
-			if desc:
-				thresh += 1
-			desc = annotator.annotate(desc)
-			self._content.append({
-				'title':title,
-				'machine_title':machine_title,
-				'description':desc,
-				'threshold':thresh,
-				'url':url
-			})
-
-	def update(self):
-		self._time = datetime.datetime.utcnow()
-		r = urllib2.urlopen("http://www.washingtonpost.com")
-		html = r.read()
-		soup = BeautifulSoup(html, "html.parser")
-		links = soup.find_all('a', {'data-pb-field':'web_headline'})
-		self.__isolate_content(links)
-		self._characterize()
 
 def uin():
 	s = input('story: ')
